requirements.txt
# Core dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# Web scraping and HTTP
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
scrapy>=2.11.0

# Search APIs
google-api-python-client>=2.108.0
bing-search>=0.1.0

# LLM Integration
anthropic>=0.7.0
openai>=1.3.0

# NLP and Text Processing
spacy>=3.7.0
nltk>=3.8.1
textblob>=0.17.1

# Data validation and serialization
xmlschema>=2.5.0
jsonschema>=4.20.0

# Metadata enrichment
isbnlib>=3.10.14
crossref-commons>=0.0.7
feedparser>=6.0.10

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.3
pydash>=7.0.6

# CLI
click>=8.1.7
rich>=13.7.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.13.0

# Async utilities
asyncio>=3.4.3
nest-asyncio>=1.5.8
